Moved AccountService to separate implementation and interface.
All tests are green.
