+++ src/main/java/rest/Sessions.java
@@ -1,6 +1,7 @@
 package rest;
 
 import main.AccountService;
+import main.AccountServiceImpl;
 import main.TokenManager;
 import org.json.JSONArray;
 import org.json.JSONException;
